+++ src/server/devices/zigbee/BaseDevices/zigbeeDimmer.ts
@@ -1,40 +1,26 @@
 import {
-  ActuatorSetStateCommand,
-  ActuatorToggleCommand,
   ActuatorWriteStateToDeviceCommand,
   BlockAutomaticCommand,
   DimmerSetLightCommand,
   DimmerSettings,
   LampSetTimeBasedCommand,
-  LampToggleLightCommand,
   LogLevel,
-  RestoreTargetAutomaticValueCommand,
 } from '../../../../models';
 import { DeviceType } from '../../deviceType';
 import { LogDebugType, SettingsService, Utils } from '../../../services';
-import { ZigbeeDevice } from './index';
 import { IoBrokerDeviceInfo } from '../../IoBrokerDeviceInfo';
 import { DeviceCapability } from '../../DeviceCapability';
 import { iDimmableLamp } from '../../baseDeviceInterfaces/iDimmableLamp';
 import { iTemporaryDisableAutomatic } from '../../baseDeviceInterfaces';
-import { BlockAutomaticHandler } from '../../../services/blockAutomaticHandler';
 import { LampUtils } from '../../sharedFunctions';
+import { ZigbeeLamp } from './zigbeeLamp';
 
-export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp, iTemporaryDisableAutomatic {
-  /** @inheritDoc */
-  public readonly blockAutomationHandler: BlockAutomaticHandler;
-  /** @inheritDoc */
-  public queuedValue: boolean | null = null;
+export abstract class ZigbeeDimmer extends ZigbeeLamp implements iDimmableLamp, iTemporaryDisableAutomatic {
   /** @inheritDoc */
   public settings: DimmerSettings = new DimmerSettings();
-  /** @inheritDoc */
-  public targetAutomaticState: boolean = false;
-  protected _lastPersist: number = 0;
   protected abstract readonly _stateIdBrightness: string;
-  protected abstract readonly _stateIdState: string;
   protected abstract readonly _stateIdTransitionTime: string;
   protected abstract readonly _stateNameBrightness: string;
-  protected abstract readonly _stateNameState: string;
   protected abstract readonly _stateNameTransitionTime: string;
 
   protected constructor(pInfo: IoBrokerDeviceInfo, deviceType: DeviceType) {
@@ -42,7 +28,6 @@ export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp
     this.deviceCapabilities.push(DeviceCapability.lamp);
     this.deviceCapabilities.push(DeviceCapability.dimmablelamp);
     this.deviceCapabilities.push(DeviceCapability.blockAutomatic);
-    this.blockAutomationHandler = new BlockAutomaticHandler(this.restoreTargetAutomaticValue.bind(this));
   }
 
   protected _brightness: number = 0;
@@ -52,24 +37,7 @@ export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp
     return this._brightness;
   }
 
-  protected _lightOn: boolean = false;
-
-  /** @inheritDoc */
-  public get lightOn(): boolean {
-    return this._lightOn;
-  }
-
   protected _transitionTime: number = 0;
-
-  /** @inheritDoc */
-  public get actuatorOn(): boolean {
-    return this.lightOn;
-  }
-
-  /** @inheritDoc */
-  public restoreTargetAutomaticValue(c: RestoreTargetAutomaticValueCommand): void {
-    this.setActuator(new ActuatorSetStateCommand(c, this.targetAutomaticState));
-  }
 
   /** @inheritDoc */
   public update(idSplit: string[], state: ioBroker.State, initial: boolean = false): void {
@@ -77,11 +45,6 @@ export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp
     this.log(LogLevel.DeepTrace, `Dimmer Update: ID: ${idSplit.join('.')} JSON: ${JSON.stringify(state)}`);
     super.update(idSplit, state, initial, true);
     switch (idSplit[3]) {
-      case this._stateNameState:
-        this.log(LogLevel.Trace, `Dimmer Update für ${this.info.customName} auf ${state.val}`);
-        this._lightOn = state.val as boolean;
-        this.persist();
-        break;
       case this._stateNameBrightness:
         this.log(LogLevel.Trace, `Dimmer Helligkeit Update für ${this.info.customName} auf ${state.val}`);
         this._brightness = state.val as number;
@@ -97,18 +60,6 @@ export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp
   /** @inheritDoc */
   public setTimeBased(c: LampSetTimeBasedCommand): void {
     this.setLight(DimmerSetLightCommand.byTimeBased(this.settings, c));
-  }
-
-  /** @inheritDoc */
-  public setActuator(c: ActuatorSetStateCommand): void {
-    this.setLight(new DimmerSetLightCommand(c, c.on, '', c.disableAutomaticCommand));
-  }
-
-  /** @inheritDoc */
-  public toggleActuator(c: ActuatorToggleCommand): boolean {
-    const setActuatorCommand: ActuatorSetStateCommand = ActuatorSetStateCommand.byActuatorAndToggleCommand(this, c);
-    this.setActuator(setActuatorCommand);
-    return setActuatorCommand.on;
   }
 
   /** @inheritDoc */
@@ -169,27 +120,6 @@ export abstract class ZigbeeDimmer extends ZigbeeDevice implements iDimmableLamp
         this,
       );
     });
-  }
-
-  /** @inheritDoc */
-  public persist(): void {
-    const now: number = Utils.nowMS();
-    if (this._lastPersist + 1000 > now) {
-      return;
-    }
-    Utils.dbo?.persistActuator(this);
-    this._lastPersist = now;
-  }
-
-  /** @inheritDoc */
-  public toggleLight(c: LampToggleLightCommand): boolean {
-    return LampUtils.toggleLight(this, c);
-  }
-
-  /** @inheritDoc */
-  public writeActuatorStateToDevice(c: ActuatorWriteStateToDeviceCommand): void {
-    this.log(LogLevel.Debug, c.logMessage, LogDebugType.SetActuator);
-    this.setState(this._stateIdState, c.stateValue);
   }
 
   private setBrightnessState(brightness: number, onSuccess?: () => void): void {