+++ src/server/devices/zigbee/zigbeeBlitzShp.ts
@@ -15,6 +15,7 @@ import { iLoadMeter } from '../baseDeviceInterfaces/iLoadMeter';
 export class ZigbeeBlitzShp extends ZigbeeActuator implements iExcessEnergyConsumer, iLoadMeter {
   /** @inheritDoc */
   public settings: ActuatorSettings = new ActuatorSettings();
+  protected readonly _stateIdState: string;
   private _steckerOn: boolean = false;
   private _current: number = 0;
   private _energy: number = 0;
@@ -23,7 +24,8 @@ export class ZigbeeBlitzShp extends ZigbeeActuator implements iExcessEnergyConsu
   private _activatedByExcessEnergy: boolean = false;
 
   public constructor(pInfo: IoBrokerDeviceInfo) {
-    super(pInfo, DeviceType.ZigbeeBlitzShp, `${pInfo.fullID}.state`);
+    super(pInfo, DeviceType.ZigbeeBlitzShp);
+    this._stateIdState = `${pInfo.fullID}.state`;
     this.settings.energySettings = new ExcessEnergyConsumerSettings();
     this.deviceCapabilities.push(DeviceCapability.excessEnergyConsumer);
     this.deviceCapabilities.push(DeviceCapability.loadMetering);