+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * (theimo1221) Correct isAutomatic detection
 * (theimo1221) Fix some inconsistent date formatting
 * (theimo1221) Fix strom-stoß devices to not block automatic on normal turn of action
+* (theimo1221) Reduce duplicate code within zigbee-actuators by introducing proper hierachy of
+  classes (`ZigbeeActuator` -> `ZigbeeLamp` -> `ZigbeeDimmer` -> `ZigbeeLedRgbcct`)
 
 ## 3.0.0